+++ browser-extension/app/components/app/app-bottom-nav.vue
