+++ browser-extension/app/components/app/app-main-button.vue
